+++ api/src/main/java/br/edu/uniceub/service/TipoAtivoService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import br.edu.uniceub.models.TipoAtivo;
 import br.edu.uniceub.repository.TipoAtivoRepository;
 import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
 import jakarta.ws.rs.Path;
 
 @Path("/tipo-ativo")
@@ -22,22 +23,32 @@ public class TipoAtivoService {
         return repository.findByIdOptional(id);
     }
 
+    @Transactional
     public TipoAtivo insereTipoAtivo(TipoAtivo novoTipoAtivo) {
         repository.persist(novoTipoAtivo);
         return novoTipoAtivo;
     }
 
+    @Transactional
     public TipoAtivo alteraTipoAtivo(Long id, TipoAtivo tipoAtivo) {
-         TipoAtivo entity = repository.findById(id);
+        TipoAtivo entity = repository.findById(id);
         if (entity != null) {
-            repository.isPersistent(tipoAtivo);
+            entity.setNomeTipoAtivo(tipoAtivo.getNomeTipoAtivo());
+            entity.setDescricao(tipoAtivo.getDescricao());
+            tipoAtivo.setId(entity.getId());
+            return tipoAtivo;
         }
-        return tipoAtivo;
+        return null;
     }
 
+    @Transactional
     public TipoAtivo deleteTipoAtivo(Long id) {
-        boolean deletado =  repository.deleteById(id);
-        return (deletado) ? repository.findById(id):null;
+        Optional<TipoAtivo> optionalTipoAtivo = repository.findByIdOptional(id);
+        if (optionalTipoAtivo.isPresent()) {
+            repository.deleteById(id);
+            return optionalTipoAtivo.get();
+        }
+        return null;
     }
-    
+
 }
